Moved chart settings functions to its own service file
